chore: remove i18n at field level
Users of our site would like to set up everything in english and then hire translators for various locales. StaticCMS requires a minimum of two locales to be declared in the i18n section of the config, so english-only would not be an option; additionally, setting any field as "i18n: true" REQUIRES translation. Removing these field declarations maintains the folder structure we need.